// chits/src/app/model/employeedetail.model.ts
export class EmployeeID {
  _id:any;
  emp_code: string;
  branch_name: string; 
  salution: string; 
  first_name: string; 
  last_name: string; 
  gender: string; 
  blood_group: string; 
  mobile_number: string; 
  email: string; 
  aadhar_no: string; 
  pan_no: string; 
  father_name: string;
  spouse_name: string; 
  designation: string; 
  qalification: string; 
  joining_date: string; 
  verified_by: string; 
  pervious_salary: string;
  bank_name: string;
  account_holder_name: string; 
  account_number: string; 
  salary: string;
  incentive: string; 
  target: string; 
  address: string; 
  pincode: string; 
  state: string; 
  distric: string; 
  city: string; 
  landmark: string; 
  experience: string; 
  ifsc_code: string; 
  remarks: string; 
  dob: string; 
}